서버 배포용 머지

+++ src/api/review/reviewControl.ts
@@ -1,4 +1,5 @@
 import { Request, Response, NextFunction } from 'express';
+import { number } from 'joi';
 import { review } from '../../interface/review';
 import reviewService from './reviewServ'; //받아온다
 
@@ -18,16 +19,15 @@ export default {
   //리뷰작성
   createReview: async (req: Request, res: Response, next: NextFunction) => {
     try {
-      // const {userId} =res.locals.user
-      const userId = 1;
+      const { userId }: review = res.locals.user;
       const { campId }: review = req.params;
       const { reviewImg, reviewComment } = req.body;
-      await reviewService.createReview(
+      await reviewService.createReview({
         userId,
-        campId!,
+        campId,
         reviewImg,
-        reviewComment
-      );
+        reviewComment,
+      });
       res.status(201).json({ ok: true, massage: '리뷰작성완료' });
     } catch (error) {
       next(error);
@@ -38,15 +38,18 @@ export default {
     try {
       const { reviewId }: review = req.params;
       const { reviewImg, reviewComment } = req.body;
-      // const { userId } = res.locals.user;
-      const userId = 1;
+      const { userId }: review = res.locals.user;
+      const findreview = await reviewService.findReviewAuthor(reviewId!);
 
-      const findreview = await reviewService.updateReview(
-        reviewId!,
+      if (userId !== findreview?.userId) {
+        return res.status(400).json({ errorMessage: '권한이 없습니다.' });
+      }
+      await reviewService.updateReview({
+        reviewId,
         reviewImg,
         reviewComment,
-        userId
-      );
+        userId,
+      });
       res.status(200).json({ massage: '리뷰수정완료' });
     } catch (error) {
       next(error);
@@ -57,33 +60,36 @@ export default {
   deleteReview: async (req: Request, res: Response, next: NextFunction) => {
     try {
       const { reviewId }: review = req.params;
-      // const { userId } = res.locals.user;
-      const userId = 1;
-      const review = await reviewService.deleteReview(reviewId!, userId);
+      const { userId } = res.locals.user;
+      const findreview = await reviewService.findReviewAuthor(reviewId!);
+
+      if (userId !== findreview?.userId) {
+        return res.status(400).json({ errorMessage: '권한이 없습니다.' });
+      }
+      await reviewService.deleteReview(reviewId!, userId);
       res.status(200).json({ massage: '리뷰삭제완료' });
     } catch (error) {
       next(error);
     }
   },
-  // //내가쓴리뷰조회
-  // getMyReview: async (req: Request, res: Response, next: NextFunction) => {
-  //   try {
-  //     // const { userId } = res.locals.user;
-  //     const userId = 1;
-  //     const myreview = await reviewService.getMyReview( userId);
 
-  //     return res.status(200).json({ data: myreview });
-  //   } catch (error) {
-  //     next(error);
-  //   }
-  // },
-  //검색하기
+  //내가쓴리뷰조회
+  getMyReview: async (req: Request, res: Response, next: NextFunction) => {
+    try {
+      const { userId } = res.locals.user;
+      const myreview = await reviewService.getMyReview(userId);
+
+      res.status(200).json({ data: myreview });
+    } catch (error) {
+      next(error);
+    }
+  },
+
+  //캠핑장이름검색
   search: async (req: Request, res: Response, next: NextFunction) => {
     try {
-      // const { userId } = res.locals.user;
-      const userId = 1;
-      const {keyword} = req.body
-      const result = await reviewService.search(userId, keyword);
+      const { keyword } = req.body;
+      const result = await reviewService.search(keyword);
 
       return res.status(200).json({ data: result });
     } catch (error) {
